refactor: get users entities
