+++ src/routes/graphql/index.ts
@@ -22,8 +22,8 @@ import { subscribeToUserTypeDefs } from './resolvers/subscribeToUser/typeDefs';
 import { subscribeToUserResolver } from './resolvers/subscribeToUser/resolver';
 import { unsubscribeFromUserTypeDefs } from './resolvers/unsubscribeFromUser/typeDefs';
 import { unsubscribeFromUserResolver } from './resolvers/unsubscribeFromUser/resolver';
-import { getUserByIdWithEntitiesResolver } from './resolvers/getUserByIdWithEntities/resolver';
-import { getUserByIdWithEntitiesTypeDefs } from './resolvers/getUserByIdWithEntities/typeDefs';
+import { getUserByIdWithEntitiesResolver } from './resolvers/getUserWithEntities/resolver';
+import { getUserByIdWithEntitiesTypeDefs } from './resolvers/getUserWithEntities/typeDefs';
 import { getUsersWithEntitiesResolver } from './resolvers/getUsersWithEntities/resolver';
 import { getUsersWithEntitiesTypeDefs } from './resolvers/getUsersWithEntities/typeDefs';
 import { getUsersProfileSubscribedToResolver } from "./resolvers/getUsersProfileSubscribedTo/resolver";
